requirements.txt
neonize
flask
python-dotenv
qrcode
Pillow